py/pyproject.toml
[tool.poetry]
name = "levenshtein-benchmark"
version = "0.1.0"
description = ""
authors = ["Miklós Koren <miklos.koren@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-Levenshtein = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
